requirements.txt
edge_tts==7.0.0
groq==0.16.0
moviepy==1.0.3
openai_whisper==20240930
Requests==2.32.3
moviepy==1.0.3
